// src/shared/api/models/getApiProductsListing200ProductsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiProductsListing200ProductsItemDescription } from './getApiProductsListing200ProductsItemDescription';
import type { GetApiProductsListing200ProductsItemOutletId } from './getApiProductsListing200ProductsItemOutletId';
import type { GetApiProductsListing200ProductsItemDeletedAt } from './getApiProductsListing200ProductsItemDeletedAt';
import type { GetApiProductsListing200ProductsItemProductsItem } from './getApiProductsListing200ProductsItemProductsItem';

export type GetApiProductsListing200ProductsItem = {
    id: string;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    seqId: number;
    name: string;
    description: GetApiProductsListing200ProductsItemDescription;
    outletId: GetApiProductsListing200ProductsItemOutletId;
    createdAt: string;
    updatedAt: string;
    deletedAt: GetApiProductsListing200ProductsItemDeletedAt;
    products: GetApiProductsListing200ProductsItemProductsItem[];
};