// src/shared/api/models/getApiProductsListing200ServicesItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiProductsListing200ServicesItemId } from './getApiProductsListing200ServicesItemId';
import type { GetApiProductsListing200ServicesItemImage } from './getApiProductsListing200ServicesItemImage';
import type { GetApiProductsListing200ServicesItemCreatedAt } from './getApiProductsListing200ServicesItemCreatedAt';
import type { GetApiProductsListing200ServicesItemUpdatedAt } from './getApiProductsListing200ServicesItemUpdatedAt';
import type { GetApiProductsListing200ServicesItemDeletedAt } from './getApiProductsListing200ServicesItemDeletedAt';
import type { GetApiProductsListing200ServicesItemServicesItem } from './getApiProductsListing200ServicesItemServicesItem';

export type GetApiProductsListing200ServicesItem = {
    id: GetApiProductsListing200ServicesItemId;
    name: string;
    image?: GetApiProductsListing200ServicesItemImage;
    needCertification?: boolean;
    outletId: number;
    createdAt: GetApiProductsListing200ServicesItemCreatedAt;
    updatedAt: GetApiProductsListing200ServicesItemUpdatedAt;
    deletedAt: GetApiProductsListing200ServicesItemDeletedAt;
    isEquipmentBased?: boolean;
    isForNewclient?: boolean;
    marketplaceHided?: boolean;
    services: GetApiProductsListing200ServicesItemServicesItem[];
};